CC-6116 Rename the property to decimal

+++ Bundles/Transfer/src/Spryker/Zed/Transfer/Business/Model/Generator/ClassDefinitionInterface.php
@@ -55,7 +55,7 @@ interface ClassDefinitionInterface extends DefinitionInterface
     /**
      * @return bool
      */
-    public function hasDouble(): bool;
+    public function hasDecimal(): bool;
 
     /**
      * @return string|null
